fixup clippy

+++ massa-db-worker/src/massa_db.rs
@@ -288,7 +288,7 @@ where
 
             for (serialized_key, serialized_value) in db_iterator.flatten() {
                 new_elements_size += serialized_value.len();
-                if new_elements_size < self.config.max_new_elements_size as usize {
+                if new_elements_size < self.config.max_new_elements_size {
                     new_elements.insert(serialized_key.to_vec(), serialized_value.to_vec());
                 } else {
                     break;
